add print_date and _time; all int8 and uint8 changed to default int; all prints back to 1st line

+++ chronometer.h
@@ -4,9 +4,9 @@
 
 class Chronometer {
     private:
-        uint8_t second;
-        uint8_t minute;
-        uint8_t hour;
+        int second;
+        int minute;
+        int hour;
         unsigned long previous_millis;
         unsigned long current_millis;
         bool is_running;
@@ -14,8 +14,8 @@ class Chronometer {
     public:
         Chronometer();
         ~Chronometer();
-        void mv_cur_up(uint8_t* pos_cursor);
-        void mv_cur_down(uint8_t* pos_cursor);
+        void mv_cur_up(int* pos_cursor);
+        void mv_cur_down(int* pos_cursor);
         void pause();
         void toggle();
         void secondCount();
